implementing quick-submit
