+++ locales/en/messages.properties
@@ -801,7 +801,7 @@ open_to_all_with_targeted_recruitment=Open to all with targeted recruitment
 random_selection=Random Selection
 election=Election
 appointment=Appointment
-undefined=undefined
+###undefined=undefined
 consultation_on_city_council_agenda=Consultation on City Council Agenda
 citizen_participation=Citizen Participation
 inuit_language=Inuit Language
@@ -1411,7 +1411,6 @@ planning_development=Planning & Development
 science_technology=Science & Technology
 social_welfare=Social Welfare
 transportation=Transportation
-
 # Specific issues
 ability/disability_issues=Ability/Disability Issues
 abortion=Abortion
@@ -1436,7 +1435,6 @@ budget_national=Budget - National
 budget_provincial_regional_state=Budget - Provincial, Regional, State
 bureaucracy=Bureaucracy
 carbon_capture_sequestration=Carbon Capture & Sequestration
-censorship=Censorship
 censorship=Censorship
 child_care=Child Care
 citizenship_role_of_citizens=Citizenship & Role of Citizens
@@ -1528,7 +1526,6 @@ land_use=Land Use
 libraries=Libraries
 longterm_care=Long-Term Care
 lowincome_assistance=Low-Income Assistance
-lowincome_assistance=Low-income Assistance
 maritime=Maritime
 mass/public_transport=Mass/Public Transport
 medical_liability=Medical Liability
@@ -1559,7 +1556,6 @@ refugee_resettlement=Refugee Resettlement
 refugee_rights=Refugee Rights
 regional_global_governance=Regional & Global Governance
 regionalism=Regionalism
-regulation=Regulation
 regulation=Regulation
 regulatory_policy=Regulatory Policy
 religious_rights=Religious Rights
@@ -1601,6 +1597,7 @@ worker_health_safety=Worker Health & Safety
 workforce_education=Workforce Education
 youth_employment=Youth Employment
 youth_issues=Youth Issues
+### Specific issues - end
 send_email_with_catgeory_additions=If you believe we should add some categories or specific topics, please email info@participedia.net
 links=Relevant Links (URLs)
 previous_page=Go to previous page
@@ -1664,7 +1661,7 @@ informal_engagement_by_intermediaries_with_political_authorities=Informal engage
 leadership_development=Leadership development (e.g., individual and group capacity for civic engagement)
 oversight_and_social_auditing=Oversight and social auditing (e.g., measuring and reporting organizational performance)
 protest=Protest (e.g., demonstrations; marches; pickets) 
-research=Research (e.g., public opinion surveys; focus groups; participatory action research)
+research_focus_groups=Research (e.g., public opinion surveys; focus groups; participatory action research)
 social_mobilization=Social mobilization (e.g., community organizing; consciousness raising; political consumerism)
 ###public_spectrum
 inform=Inform (provide the public with balanced and objective information to assist them in understanding the problem, alternatives, opportunities, and/or solutions)
