+++ src/components/choices.js
@@ -1,7 +1,7 @@
 const caseChoiceData = require("./case.choices.json");
 const methodChoiceData = require("./method.choices.json");
 const orgChoiceData = require("./organization.choices.json");
-const issueChoiceData = require("./issuemap.json");
+// const issueChoiceData = require("./issuemap.json");
 
 let choiceData = {};
 for (let key in caseChoiceData) {
@@ -16,10 +16,10 @@ for (let key in orgChoiceData) {
   if (!orgChoiceData.hasOwnProperty(key)) continue;
   choiceData[key] = orgChoiceData[key];
 }
-for (let key in issueChoiceData) {
-  if (!issueChoiceData.hasOwnProperty(key)) continue;
-  choiceData[key] = issueChoiceData[key];
-}
+// for (let key in issueChoiceData) {
+//   if (!issueChoiceData.hasOwnProperty(key)) continue;
+//   choiceData[key] = issueChoiceData[key];
+// }
 
 export function getChoices(property) {
   if (choiceData[property]) {
@@ -38,7 +38,7 @@ function toTitleCase(str) {
   }
 }
 
-export function makeLocalizedChoices(intl, property) {
+export function makeLocalizedChoices(intl, property, alphabetical = true) {
   let hasOther;
   let choices = getChoices(property).map(function(v) {
     return {
@@ -46,36 +46,38 @@ export function makeLocalizedChoices(intl, property) {
       value: v
     };
   });
-  // choices.sort(function(a, b) {
-  //   if (typeof a === undefined) return -1;
-  //   if (typeof b === undefined) return 1;
-  //   if (!a.text) return -1;
-  //   if (!b.text) return 1;
-  //   if (typeof a === typeof {}) {
-  //     if (a.text.toLowerCase() < b.text.toLowerCase()) return -1;
-  //     if (a.text.toLowerCase() === b.text.toLowerCase()) return 0;
-  //   } else {
-  //     if (a < b) {
-  //       return -1;
-  //     }
-  //     if (a === b) {
-  //       return 0;
-  //     }
-  //   }
-  //   return 1;
-  // });
+  if (alphabetical) {
+    choices.sort(function(a, b) {
+      if (typeof a === undefined) return -1;
+      if (typeof b === undefined) return 1;
+      if (!a.text) return -1;
+      if (!b.text) return 1;
+      if (typeof a === typeof {}) {
+        if (a.text.toLowerCase() < b.text.toLowerCase()) return -1;
+        if (a.text.toLowerCase() === b.text.toLowerCase()) return 0;
+      } else {
+        if (a < b) {
+          return -1;
+        }
+        if (a === b) {
+          return 0;
+        }
+      }
+      return 1;
+    });
 
-  // Object.keys(choices).forEach(function(key) {
-  //   if (choices[key]['text'] === "Other") {
-  //     hasOther = true;
-  //     delete choices[key];
-  //   }
-  // }); 
+    Object.keys(choices).forEach(function(key) {
+      if (choices[key]['text'] === "Other") {
+        hasOther = true;
+        delete choices[key];
+      }
+    }); 
 
-  // if (hasOther) {
-  //   let choicesLength = choices.length
-  //   choices[choicesLength] = { value: "other", text: "Other" };
-  // };
+    if (hasOther) {
+      let choicesLength = choices.length
+      choices[choicesLength] = { value: "other", text: "Other" };
+    };
+  }
   
   return choices
