+++ src/locales.json
@@ -7,7 +7,7 @@
     "chronological": "Most Recent",
     "alphabetical": "Alphabetical",
     "about": "About",
-    "research": "Research (e.g., public opinion surveys; focus groups; participatory action research)",
+    "research": "Research",
     "all": "All",
     "news": "News",
     "case": "Case",
@@ -744,7 +744,6 @@
     "random_selection": "Random Selection",
     "election": "Election",
     "appointment": "Appointment",
-    "undefined": "undefined",
     "consultation_on_city_council_agenda": "Consultation on City Council Agenda",
     "citizen_participation": "Citizen Participation",
     "inuit_language": "Inuit Language",
@@ -1411,7 +1410,7 @@
     "land_use": "Land Use",
     "libraries": "Libraries",
     "longterm_care": "Long-Term Care",
-    "lowincome_assistance": "Low-income Assistance",
+    "lowincome_assistance": "Low-Income Assistance",
     "maritime": "Maritime",
     "mass/public_transport": "Mass/Public Transport",
     "medical_liability": "Medical Liability",
@@ -1529,6 +1528,7 @@
     "informal_engagement_by_intermediaries_with_political_authorities": "Informal engagement by intermediaries with political authorities (e.g., bargaining and negotiation on behalf of marginalized sectors )",
     "leadership_development": "Leadership development (e.g., individual and group capacity for civic engagement)",
     "oversight_and_social_auditing": "Oversight and social auditing (e.g., measuring and reporting organizational performance)",
+    "research_focus_groups": "Research (e.g., public opinion surveys; focus groups; participatory action research)",
     "social_mobilization": "Social mobilization (e.g., community organizing; consciousness raising; political consumerism)",
     "inform": "Inform (provide the public with balanced and objective information to assist them in understanding the problem, alternatives, opportunities, and/or solutions)",
     "consult": "Consult (obtain public feedback on analyses, alternatives, and/or decisions)",
